Move screens to packages.
